Refactor utils tests to use ginkgo/gomega
